# MAINT: simplify power fast path logic