--- a/numpy/_core/src/multiarray/number.c
+++ b/numpy/_core/src/multiarray/number.c
@@ -328,55 +328,6 @@ array_inplace_matrix_multiply(PyArrayObject *self, PyObject *other)
     return res;
 }
 
-static int
-fast_scalar_power(PyObject *o1, PyObject *o2, int inplace, PyObject **result)
-{
-    PyObject *fastop = NULL;
-    if (PyLong_CheckExact(o2)) {
-        int overflow = 0;
-        long exp = PyLong_AsLongAndOverflow(o2, &overflow);
-        if (overflow != 0) {
-            return -1;
-        }
-
-        if (exp == -1) {
-            fastop = n_ops.reciprocal;
-        }
-        else if (exp == 2) {
-            fastop = n_ops.square;
-        }
-        else {
-            return 1;
-        }
-    }
-    else if (PyFloat_CheckExact(o2)) {
-        double exp = PyFloat_AsDouble(o2);
-        if (exp == 0.5) {
-            fastop = n_ops.sqrt;
-        }
-        else {
-            return 1;
-        }
-    }
-    else {
-        return 1;
-    }
-
-    PyArrayObject *a1 = (PyArrayObject *)o1;
-    if (PyArray_ISOBJECT(a1)) {
-        return 1;
-    }
-
-    if (inplace || can_elide_temp_unary(a1)) {
-        *result = PyArray_GenericInplaceUnaryFunction(a1, fastop);
-    }
-    else {
-        *result = PyArray_GenericUnaryFunction(a1, fastop);
-    }
-
-    return 0;
-}
-
 static PyObject *
 array_power(PyObject *a1, PyObject *o2, PyObject *modulo)
 {
@@ -389,9 +340,7 @@ array_power(PyObject *a1, PyObject *o2, PyObject *modulo)
     }
 
     BINOP_GIVE_UP_IF_NEEDED(a1, o2, nb_power, array_power);
-    if (fast_scalar_power(a1, o2, 0, &value) != 0) {
-        value = PyArray_GenericBinaryFunction(a1, o2, n_ops.power);
-    }
+    value = PyArray_GenericBinaryFunction(a1, o2, n_ops.power);
     return value;
 }
 
@@ -531,9 +480,7 @@ array_inplace_power(PyArrayObject *a1, PyObject *o2, PyObject *NPY_UNUSED(modulo
 
     INPLACE_GIVE_UP_IF_NEEDED(
             a1, o2, nb_inplace_power, array_inplace_power);
-    if (fast_scalar_power((PyObject *) a1, o2, 1, &value) != 0) {
-        value = PyArray_GenericInplaceBinaryFunction(a1, o2, n_ops.power);
-    }
+    value = PyArray_GenericInplaceBinaryFunction(a1, o2, n_ops.power);
     return value;
 }