--- a/numpy/_core/src/multiarray/number.c
+++ b/numpy/_core/src/multiarray/number.c
@@ -363,7 +363,7 @@ fast_scalar_power(PyObject *o1, PyObject *o2, int inplace, PyObject **result)
     }
 
     PyArrayObject *a1 = (PyArrayObject *)o1;
-    if (PyArray_ISOBJECT(a1)) {
+    if (!(PyArray_ISFLOAT(a1) || PyArray_ISCOMPLEX(a1))) {
         return 1;
     }