--- a/numpy/_core/src/umath/loops.c.src
+++ b/numpy/_core/src/umath/loops.c.src
@@ -525,15 +525,25 @@ NPY_NO_EXPORT void
         if (in2 == 0) {
             *((@type@ *)op1) = 1;
             continue;
+        } 
+        else if (in2 == 1)
+        {
+            *((@type@ *)op1) = *(@type@ *)ip1;
         }
-        if (in1 == 1) {
+        else if (in2 == 2)
+        {
+            *((@type@ *)op1) = *(@type@ *)ip1 * *(@type@ *)ip1;
+        }
+        else if (in1 == 1) {
             *((@type@ *)op1) = 1;
             continue;
         }
-
-        int first_bit = in2 & 1;
-        in2 >>= 1;
-        *((@type@ *) op1) = _@TYPE@_squared_exponentiation_helper(in1, in2, first_bit);
+        else
+        {
+            int first_bit = in2 & 1;
+            in2 >>= 1;
+            *((@type@ *) op1) = _@TYPE@_squared_exponentiation_helper(in1, in2, first_bit);
+        }
     }
 }
 /**end repeat**/