--- a/numpy/_core/src/umath/loops.c.src
+++ b/numpy/_core/src/umath/loops.c.src
@@ -524,19 +524,17 @@ NPY_NO_EXPORT void
 #endif
         if (in2 == 0) {
             *((@type@ *)op1) = 1;
-            continue;
         } 
         else if (in2 == 1)
         {
-            *((@type@ *)op1) = *(@type@ *)ip1;
+            *((@type@ *)op1) = in1;
         }
         else if (in2 == 2)
         {
-            *((@type@ *)op1) = *(@type@ *)ip1 * *(@type@ *)ip1;
+            *((@type@ *)op1) = in1 * in1;
         }
         else if (in1 == 1) {
             *((@type@ *)op1) = 1;
-            continue;
         }
         else
         {