--- a/numpy/_core/src/umath/loops.c.src
+++ b/numpy/_core/src/umath/loops.c.src
@@ -486,6 +486,35 @@ _@TYPE@_squared_exponentiation_helper(@type@ base, @type@ exponent_two, int firs
    return out;
 }
 
+static inline @type@
+_@TYPE@_power_fast_path_helper(@type@ in1, @type@ in2, @type@ *op1) {
+    // Fast path for power calculation
+    #if @SIGNED@
+        if (in2 < 0) {
+            npy_gil_error(PyExc_ValueError,
+                            "Integers to negative integer powers are not allowed.");
+            return;
+        }
+    #endif
+    if (in1 == 1) {
+        *op1 = 1;
+    }
+    else if (in2 == 0 & in1 != 0) {
+        *op1 = 1;
+    }
+    else if (in2 == 1) {
+        *op1 = in1;
+    }
+    else if (in2 == 2) {
+        *op1 = in1 * in1;
+    }
+    else {
+        return 1;
+    }
+    return 0;
+}
+
+
 NPY_NO_EXPORT void
 @TYPE@_power(char **args, npy_intp const *dimensions, npy_intp const *steps, void *NPY_UNUSED(func))
 {
@@ -507,7 +536,9 @@ NPY_NO_EXPORT void
         BINARY_LOOP_SLIDING {
             @type@ in1 = *(@type@ *)ip1;
 
-            *((@type@ *) op1) = _@TYPE@_squared_exponentiation_helper(in1, in2start, first_bit);
+            if (_@TYPE@_power_fast_path_helper(in1, in2start, (@type@ *)op1) != 0) {
+                *((@type@ *) op1) = _@TYPE@_squared_exponentiation_helper(in1, in2start, first_bit);
+            }
         }
         return;
     }
@@ -515,29 +546,7 @@ NPY_NO_EXPORT void
         @type@ in1 = *(@type@ *)ip1;
         @type@ in2 = *(@type@ *)ip2;
 
-#if @SIGNED@
-        if (in2 < 0) {
-            npy_gil_error(PyExc_ValueError,
-                          "Integers to negative integer powers are not allowed.");
-            return;
-        }
-#endif
-        if (in2 == 0) {
-            *((@type@ *)op1) = 1;
-        } 
-        else if (in2 == 1)
-        {
-            *((@type@ *)op1) = in1;
-        }
-        else if (in2 == 2)
-        {
-            *((@type@ *)op1) = in1 * in1;
-        }
-        else if (in1 == 1) {
-            *((@type@ *)op1) = 1;
-        }
-        else
-        {
+        if (_@TYPE@_power_fast_path_helper(in1, in2, (@type@ *)op1) != 0) {
             int first_bit = in2 & 1;
             in2 >>= 1;
             *((@type@ *) op1) = _@TYPE@_squared_exponentiation_helper(in1, in2, first_bit);