--- a/numpy/_core/src/umath/loops.c.src
+++ b/numpy/_core/src/umath/loops.c.src
@@ -523,14 +523,13 @@ NPY_NO_EXPORT void
         BINARY_DEFS
         const @type@ in2 = *(@type@ *)ip2;
 
-        @type@ in2_working = in2;
-        int first_bit = in2_working & 1;
-        in2_working >>= 1;
+        int first_bit = in2 & 1;
+        @type@ in2start = in2 >> 1;
 
         BINARY_LOOP_SLIDING {
             @type@ in1 = *(@type@ *)ip1;
             if (_@TYPE@_power_fast_path_helper(in1, in2, (@type@ *)op1) != 0) {
-                *((@type@ *) op1) = _@TYPE@_squared_exponentiation_helper(in1, in2_working, first_bit);
+                *((@type@ *) op1) = _@TYPE@_squared_exponentiation_helper(in1, in2start, first_bit);
             }
         }
         return;
@@ -538,12 +537,11 @@ NPY_NO_EXPORT void
     BINARY_LOOP {
         @type@ in1 = *(@type@ *)ip1;
         @type@ in2 = *(@type@ *)ip2;
-        @type@ in2_working = in2;
 
         if (_@TYPE@_power_fast_path_helper(in1, in2, (@type@ *)op1) != 0) {
-            int first_bit = in2_working & 1;
-            in2_working >>= 1;
-            *((@type@ *) op1) = _@TYPE@_squared_exponentiation_helper(in1, in2_working, first_bit);
+            int first_bit = in2 & 1;
+            in2 >>= 1;
+            *((@type@ *) op1) = _@TYPE@_squared_exponentiation_helper(in1, in2, first_bit);
         }
     }
 }