--- a/numpy/_core/src/umath/loops.c.src
+++ b/numpy/_core/src/umath/loops.c.src
@@ -489,13 +489,6 @@ _@TYPE@_squared_exponentiation_helper(@type@ base, @type@ exponent_two, int firs
 static inline @type@
 _@TYPE@_power_fast_path_helper(@type@ in1, @type@ in2, @type@ *op1) {
     // Fast path for power calculation
-    #if @SIGNED@
-        if (in2 < 0) {
-            npy_gil_error(PyExc_ValueError,
-                            "Integers to negative integer powers are not allowed.");
-            return 0;
-        }
-    #endif
     if (in1 == 1) {
         *op1 = 1;
     }
@@ -523,12 +516,28 @@ NPY_NO_EXPORT void
         BINARY_DEFS
         const @type@ in2 = *(@type@ *)ip2;
 
+#if @SIGNED@
+        if (in2 < 0) {
+            npy_gil_error(PyExc_ValueError,
+                            "Integers to negative integer powers are not allowed.");
+            return;
+        }
+#endif
+
         int first_bit = in2 & 1;
         @type@ in2start = in2 >> 1;
 
-        BINARY_LOOP_SLIDING {
-            @type@ in1 = *(@type@ *)ip1;
-            if (_@TYPE@_power_fast_path_helper(in1, in2, (@type@ *)op1) != 0) {
+        if (in2 == 0 || in2 == 1 || in2 == 2) {
+            BINARY_LOOP {
+                @type@ in1 = *(@type@ *)ip1;
+                if (_@TYPE@_power_fast_path_helper(in1, in2, (@type@ *)op1) != 0) {
+                    *((@type@ *) op1) = _@TYPE@_squared_exponentiation_helper(in1, in2start, first_bit);
+                }
+            }
+        }
+        else {
+            BINARY_LOOP_SLIDING {
+                @type@ in1 = *(@type@ *)ip1;
                 *((@type@ *) op1) = _@TYPE@_squared_exponentiation_helper(in1, in2start, first_bit);
             }
         }
@@ -537,6 +546,14 @@ NPY_NO_EXPORT void
     BINARY_LOOP {
         @type@ in1 = *(@type@ *)ip1;
         @type@ in2 = *(@type@ *)ip2;
+
+#if @SIGNED@
+        if (in2 < 0) {
+            npy_gil_error(PyExc_ValueError,
+                            "Integers to negative integer powers are not allowed.");
+            return;
+        }
+#endif
 
         if (_@TYPE@_power_fast_path_helper(in1, in2, (@type@ *)op1) != 0) {
             int first_bit = in2 & 1;