… numpy/_core/src/umath/loops_umath_fp.dispatch.c.src
… numpy/_core/tests/test_multiarray.py
… numpy/linalg/tests/test_regression.py