@seberg 2024-12-04
Can we (I suppose we can) trust that the compiler is smart enough to lift that `if (step[1] == 0)` out of the loop?
Either way, I don't really think a 5% slowdown is super concerning, and it may also be partially a fluke.

The main difference should be that for `**2` the old code didn't have to create `np.array(2)` to do the operation.  So adding these fast-paths sacrifices that.

Not sure we should be concerned about it, but if anyone is, an *exact* fast path for common Python integers only to call `square` or `reciprocal` could still make sense.
(Exact fast path for integers, because that is the normal code and makes the fast-path code very simple.)