@MaanasArora 2024-12-04
Yes, a surface-level check in `array_power` for Python integers and floats (exact type matches) should be a good trade-off, so that operators don't have to cast to numpy. Going to implement that, and the float fast paths. Thanks.