@MaanasArora 2024-12-04
Both float fast paths and the array power fast paths are complete!

It seems a specific test under `test_regression` is breaking. As far as I can tell, the issue seems to be that, because of calling `reciprocal` on the object arrays during execution of `linalg.norm`, there is a `RuntimeWarning` (zero division) instead of a `ValueError`. On inspecting the other occurrences, it doesn't seem like the other `ValueError`s are exactly the same either, so I'm assuming I can change the test to `RuntimeWarning`. It does seem inconsistent, but I suppose it indicates a problem with the test logic?