@MaanasArora 2024-12-05
The regression test seems to concern norms on object arrays. I have updated the test to consider a larger amount of possible exceptions, though I am not sure if I made it too permissive (perhaps it should only be done for -1 norms)? Regardless, if someone was doing -1 norms on object arrays which contain zeros, I doubt they would be surprised to see a `ZeroDivisionError` or similar.