@MaanasArora 2024-12-05
My bad, just fixed an erroneously declared extra variable in the fast paths.