@MaanasArora 2024-12-05
Wrote an enhancing check to improve speed when the exponent is a scalar.

Here are the benchmarks with changes:

```[57.14%] ··· bench_ufunc.BinaryBench.time_atan2                                                                                            ok
[57.14%] ··· =============== ============
                  dtype                  
             --------------- ------------
              numpy.float32   23.7±0.1ms 
              numpy.float64   19.4±0.1ms 
             =============== ============

[64.29%] ··· bench_ufunc.BinaryBench.time_pow                                                                                              ok
[64.29%] ··· =============== =============
                  dtype                   
             --------------- -------------
              numpy.float32   7.40±0.04ms 
              numpy.float64   12.0±0.07ms 
             =============== =============

[71.43%] ··· bench_ufunc.BinaryBench.time_pow_2                                                                                            ok
[71.43%] ··· =============== =============
                  dtype                   
             --------------- -------------
              numpy.float32   1.97±0.01ms 
              numpy.float64   2.14±0.01ms 
             =============== =============

[78.57%] ··· bench_ufunc.BinaryBench.time_pow_half                                                                                         ok
[78.57%] ··· =============== =============
                  dtype                   
             --------------- -------------
              numpy.float32   2.39±0.03ms 
              numpy.float64    2.54±0.3ms 
             =============== =============

[85.71%] ··· bench_ufunc.BinaryBenchInteger.time_pow                                                                                       ok
[85.71%] ··· ============= =============
                 dtype                  
             ------------- -------------
              numpy.int32   10.2±0.06ms 
              numpy.int64   10.2±0.07ms 
             ============= =============

[92.86%] ··· bench_ufunc.BinaryBenchInteger.time_pow_five                                                                                  ok
[92.86%] ··· ============= =============
                 dtype                  
             ------------- -------------
              numpy.int32   3.66±0.02ms 
              numpy.int64    3.93±0.5ms 
             ============= =============

[100.00%] ··· bench_ufunc.BinaryBenchInteger.time_pow_two                                                                                   ok
[100.00%] ··· ============= =============
                  dtype                  
              ------------- -------------
               numpy.int32   3.12±0.01ms 
               numpy.int64    3.37±0.2ms 
              ============= =============
```

And without:

```[57.14%] ··· bench_ufunc.BinaryBench.time_atan2                                                                                            ok
[57.14%] ··· =============== =============
                  dtype                   
             --------------- -------------
              numpy.float32   23.6±0.02ms 
              numpy.float64   19.3±0.03ms 
             =============== =============

[64.29%] ··· bench_ufunc.BinaryBench.time_pow                                                                                              ok
[64.29%] ··· =============== =============
                  dtype                   
             --------------- -------------
              numpy.float32   7.43±0.02ms 
              numpy.float64     11.9±0ms  
             =============== =============

[71.43%] ··· bench_ufunc.BinaryBench.time_pow_2                                                                                            ok
[71.43%] ··· =============== =============
                  dtype                   
             --------------- -------------
              numpy.float32   2.04±0.01ms 
              numpy.float64   1.99±0.01ms 
             =============== =============

[78.57%] ··· bench_ufunc.BinaryBench.time_pow_half                                                                                         ok
[78.57%] ··· =============== =============
                  dtype                   
             --------------- -------------
              numpy.float32   7.41±0.01ms 
              numpy.float64    12.1±0.2ms 
             =============== =============

[85.71%] ··· bench_ufunc.BinaryBenchInteger.time_pow                                                                                       ok
[85.71%] ··· ============= =============
                 dtype                  
             ------------- -------------
              numpy.int32   9.53±0.06ms 
              numpy.int64    9.74±0.3ms 
             ============= =============

[92.86%] ··· bench_ufunc.BinaryBenchInteger.time_pow_five                                                                                  ok
[92.86%] ··· ============= =============
                 dtype                  
             ------------- -------------
              numpy.int32   3.88±0.02ms 
              numpy.int64    3.97±0.2ms 
             ============= =============

[100.00%] ··· bench_ufunc.BinaryBenchInteger.time_pow_two                                                                                   ok
[100.00%] ··· ============= =============
                  dtype                  
              ------------- -------------
               numpy.int32   3.28±0.01ms 
               numpy.int64   3.49±0.01ms 
              ============= =============
```