@MaanasArora 2024-12-05
No, don't mind at all! I appreciate the feedback.

I am rethinking some of the logic I built to pass the symmetry checks. Perhaps it will help clean up the code further.