@MaanasArora 2024-12-10
After running the benchmarks, it seems that the performance decreases were indeed flukes, and on the whole, checking for Python scalars might have actually reduced performance by a slight amount. I removed them to decrease the amount of code. If we would still like to have them, happy to add back.

I couldn't find an efficient way to dispatch `@TYPE@_sqrt` and the others; they required moving around the memory. To be fair, these are in the UFuncs for specific types.

Here are the compared benchmarks (run with `asv run -a repeat=50 -a rounds=10` to reduce noise):

```
| Change   | Before [5047f7be] <master>   | After [b5f8a2be] <maint/simpler-power-fast-paths>   |   Ratio | Benchmark (Parameter)                                               |
|----------|------------------------------|-----------------------------------------------------|---------|---------------------------------------------------------------------|
|          | 17.9±0.3ms                   | 17.7±0.09ms                                         |    0.99 | bench_ufunc.BinaryBench.time_atan2(<class 'numpy.float32'>)         |
|          | 14.9±0.1ms                   | 14.8±0.1ms                                          |    0.99 | bench_ufunc.BinaryBench.time_atan2(<class 'numpy.float64'>)         |
|          | 4.93±0.2ms                   | 4.96±0.06ms                                         |    1.01 | bench_ufunc.BinaryBench.time_pow(<class 'numpy.float32'>)           |
|          | 10.5±0.5ms                   | 10.5±0.07ms                                         |    1    | bench_ufunc.BinaryBench.time_pow(<class 'numpy.float64'>)           |
|          | 493±30μs                     | 540±20μs                                            |    1.1  | bench_ufunc.BinaryBench.time_pow_2(<class 'numpy.float32'>)         |
| -        | 937±200μs                    | 581±30μs                                            |    0.62 | bench_ufunc.BinaryBench.time_pow_2(<class 'numpy.float64'>)         |
| -        | 4.94±0.04ms                  | 802±6μs                                             |    0.16 | bench_ufunc.BinaryBench.time_pow_half(<class 'numpy.float32'>)      |
| -        | 10.4±0.06ms                  | 1.61±0.02ms                                         |    0.15 | bench_ufunc.BinaryBench.time_pow_half(<class 'numpy.float64'>)      |
| +        | 5.27±0.1ms                   | 6.85±0.04ms                                         |    1.3  | bench_ufunc.BinaryBenchInteger.time_pow(<class 'numpy.int32'>)      |
| +        | 5.56±0.2ms                   | 6.92±0.1ms                                          |    1.25 | bench_ufunc.BinaryBenchInteger.time_pow(<class 'numpy.int64'>)      |
|          | 1.25±0.04ms                  | 1.33±0.01ms                                         |    1.07 | bench_ufunc.BinaryBenchInteger.time_pow_five(<class 'numpy.int32'>) |
|          | 1.42±0.2ms                   | 1.36±0.04ms                                         |    0.96 | bench_ufunc.BinaryBenchInteger.time_pow_five(<class 'numpy.int64'>) |
| +        | 846±50μs                     | 1.16±0.03ms                                         |    1.38 | bench_ufunc.BinaryBenchInteger.time_pow_two(<class 'numpy.int32'>)  |
|          | 1.15±0.2ms                   | 1.17±0.1ms                                          |    1.01 | bench_ufunc.BinaryBenchInteger.time_pow_two(<class 'numpy.int64'>)  |
```

This should be ready for review! Thank you.