@MaanasArora 2024-12-11
Thanks! That makes sense. Added the benchmarks. Yes, float operators did get slower:

<details>

```
| Change   | Before [5047f7be] <master>   | After [b5f8a2be] <maint/simpler-power-fast-paths>   |   Ratio | Benchmark (Parameter)                                               |
|----------|------------------------------|-----------------------------------------------------|---------|---------------------------------------------------------------------|
|          | 17.7±0.05ms                  | 17.9±0.4ms                                          |    1.01 | bench_ufunc.BinaryBench.time_atan2(<class 'numpy.float32'>)         |
|          | 14.9±0.2ms                   | 15.0±0.09ms                                         |    1.01 | bench_ufunc.BinaryBench.time_atan2(<class 'numpy.float64'>)         |
|          | 4.92±0.05ms                  | 5.01±0.09ms                                         |    1.02 | bench_ufunc.BinaryBench.time_pow(<class 'numpy.float32'>)           |
|          | 10.5±0.2ms                   | 10.7±0.3ms                                          |    1.01 | bench_ufunc.BinaryBench.time_pow(<class 'numpy.float64'>)           |
| +        | 419±5μs                      | 629±60μs                                            |    1.5  | bench_ufunc.BinaryBench.time_pow_2(<class 'numpy.float32'>)         |
| +        | 472±50μs                     | 1.16±0.5ms                                          |    2.45 | bench_ufunc.BinaryBench.time_pow_2(<class 'numpy.float64'>)         |
v
| +        | 165±8μs                      | 596±70μs                                            |    3.61 | bench_ufunc.BinaryBench.time_pow_2_op(<class 'numpy.float32'>)      |
| +        | 344±5μs                      | 1.04±0.3ms                                          |    3.01 | bench_ufunc.BinaryBench.time_pow_2_op(<class 'numpy.float64'>)      |
^
| -        | 4.88±0.04ms                  | 893±50μs                                            |    0.18 | bench_ufunc.BinaryBench.time_pow_half(<class 'numpy.float32'>)      |
| -        | 10.3±0.05ms                  | 2.19±0.2ms                                          |    0.21 | bench_ufunc.BinaryBench.time_pow_half(<class 'numpy.float64'>)      |
v
| +        | 201±0.4μs                    | 915±40μs                                            |    4.55 | bench_ufunc.BinaryBench.time_pow_half_op(<class 'numpy.float32'>)   |
| +        | 806±2μs                      | 1.96±0.3ms                                          |    2.43 | bench_ufunc.BinaryBench.time_pow_half_op(<class 'numpy.float64'>)   |
^
| +        | 5.12±0.02ms                  | 7.44±0.3ms                                          |    1.45 | bench_ufunc.BinaryBenchInteger.time_pow(<class 'numpy.int32'>)      |
| +        | 5.35±0.07ms                  | 7.02±0.3ms                                          |    1.31 | bench_ufunc.BinaryBenchInteger.time_pow(<class 'numpy.int64'>)      |
| +        | 1.20±0ms                     | 1.50±0.07ms                                         |    1.25 | bench_ufunc.BinaryBenchInteger.time_pow_five(<class 'numpy.int32'>) |
| +        | 1.25±0.02ms                  | 1.70±0.3ms                                          |    1.35 | bench_ufunc.BinaryBenchInteger.time_pow_five(<class 'numpy.int64'>) |
| +        | 802±2μs                      | 1.31±0.1ms                                          |    1.63 | bench_ufunc.BinaryBenchInteger.time_pow_two(<class 'numpy.int32'>)  |
| +        | 823±20μs                     | 1.55±0.5ms                                          |    1.88 | bench_ufunc.BinaryBenchInteger.time_pow_two(<class 'numpy.int64'>)  |
```

</details>

I've added the fast paths back:

<details>

```
| Change   | Before [5047f7be] <master>   | After [efbd6b83] <maint/simpler-power-fast-paths>   | Ratio   | Benchmark (Parameter)                                               |
|----------|------------------------------|-----------------------------------------------------|---------|---------------------------------------------------------------------|
|          | 17.9±0.4ms                   | 17.9±0.05ms                                         | 1.00    | bench_ufunc.BinaryBench.time_atan2(<class 'numpy.float32'>)         |
|          | 15.0±0.5ms                   | 14.7±0.2ms                                          | 0.98    | bench_ufunc.BinaryBench.time_atan2(<class 'numpy.float64'>)         |
|          | 4.99±0.1ms                   | 4.97±0.04ms                                         | 1.00    | bench_ufunc.BinaryBench.time_pow(<class 'numpy.float32'>)           |
|          | 10.4±0.1ms                   | 10.4±0.07ms                                         | 1.01    | bench_ufunc.BinaryBench.time_pow(<class 'numpy.float64'>)           |
|          | 500±30μs                     | 534±2μs                                             | 1.07    | bench_ufunc.BinaryBench.time_pow_2(<class 'numpy.float32'>)         |
|          | 836±500μs                    | 557±30μs                                            | ~0.67   | bench_ufunc.BinaryBench.time_pow_2(<class 'numpy.float64'>)         |
|          | 210±50μs                     | 167±2μs                                             | ~0.79   | bench_ufunc.BinaryBench.time_pow_2_op(<class 'numpy.float32'>)      |
|          | 404±100μs                    | 347±20μs                                            | ~0.86   | bench_ufunc.BinaryBench.time_pow_2_op(<class 'numpy.float64'>)      |
| -        | 4.96±0.04ms                  | 797±0.9μs                                           | 0.16    | bench_ufunc.BinaryBench.time_pow_half(<class 'numpy.float32'>)      |
| -        | 10.6±0.1ms                   | 1.60±0ms                                            | 0.15    | bench_ufunc.BinaryBench.time_pow_half(<class 'numpy.float64'>)      |
|          | 202±10μs                     | 201±0.6μs                                           | 1.00    | bench_ufunc.BinaryBench.time_pow_half_op(<class 'numpy.float32'>)   |
|          | 811±50μs                     | 806±2μs                                             | 0.99    | bench_ufunc.BinaryBench.time_pow_half_op(<class 'numpy.float64'>)   |
| +        | 5.18±0.1ms                   | 6.82±0.02ms                                         | 1.32    | bench_ufunc.BinaryBenchInteger.time_pow(<class 'numpy.int32'>)      |
| +        | 5.39±0.2ms                   | 6.97±0.1ms                                          | 1.29    | bench_ufunc.BinaryBenchInteger.time_pow(<class 'numpy.int64'>)      |
|          | 1.19±0.04ms                  | 1.33±0ms                                            | ~1.11   | bench_ufunc.BinaryBenchInteger.time_pow_five(<class 'numpy.int32'>) |
|          | 1.26±0.1ms                   | 1.35±0.01ms                                         | 1.07    | bench_ufunc.BinaryBenchInteger.time_pow_five(<class 'numpy.int64'>) |
| +        | 797±20μs                     | 1.13±0.02ms                                         | 1.42    | bench_ufunc.BinaryBenchInteger.time_pow_two(<class 'numpy.int32'>)  |
|          | 895±200μs                    | 1.14±0.01ms                                         | ~1.28   | bench_ufunc.BinaryBenchInteger.time_pow_two(<class 'numpy.int64'>)  |
```

</details>

The `RuntimeWarning` case, which occurs when `ord = -1`, does eventually trigger a `ZeroDivisionError` as well:

<details>

```
In [1]: from numpy import linalg

In [2]: testvector = np.array([np.array([0, 1]), 0, 0], dtype=object)

In [3]: linalg.norm(testvector, ord=-1)
/home/maanas/Documents/open-source/numpy/build-install/usr/lib/python3.12/site-packages/numpy/linalg/_linalg.py:2788: RuntimeWarning: divide by zero encountered in reciprocal
  absx **= ord
/home/maanas/Documents/open-source/numpy/build-install/usr/lib/python3.12/site-packages/numpy/linalg/_linalg.py:2788: RuntimeWarning: invalid value encountered in reciprocal
  absx **= ord
---------------------------------------------------------------------------
ZeroDivisionError                         Traceback (most recent call last)
Cell In[3], line 1
----> 1 linalg.norm(testvector, ord=-1)

File ~/Documents/open-source/numpy/build-install/usr/lib/python3.12/site-packages/numpy/linalg/_linalg.py:2788, in norm(x, ord, axis, keepdims)
   2786 else:
   2787     absx = abs(x)
-> 2788     absx **= ord
   2789     ret = add.reduce(absx, axis=axis, keepdims=keepdims)
   2790     ret **= reciprocal(ord, dtype=ret.dtype)

ZeroDivisionError: 0.0 cannot be raised to a negative power
```

</details>

If that looks fine, I could look into adjusting the test for sure, to ensure the warning does trigger one of the errors!